+++ tasmota/i18n.h
@@ -52,6 +52,7 @@
 #define D_JSON_CURRENT "Current"         // As in Voltage and Current
 #define D_JSON_DARKNESS "Darkness"
 #define D_JSON_DATA "Data"
+#define D_JSON_DEWPOINT "DewPoint"
 #define D_JSON_DISTANCE "Distance"
 #define D_JSON_DNSSERVER "DNSServer"
 #define D_JSON_DONE "Done"
@@ -62,6 +63,7 @@
 #define D_JSON_ENERGY "Energy"
 #define D_JSON_ERASE "Erase"
 #define D_JSON_ERROR "Error"
+#define D_JSON_EVENT "Event"
 #define D_JSON_EVERY "Every"
 #define D_JSON_EXPORT_ACTIVE "ExportActive"
 #define D_JSON_EXPORT_REACTIVE "ExportReactive"
@@ -139,6 +141,8 @@
 #define D_JSON_SERIALRECEIVED "SerialReceived"
 #define D_JSON_SET "Set"
 #define D_JSON_SIGNAL "Signal"
+#define D_JSON_SPEED "Speed"
+#define D_JSON_SPEED_UNIT "SpeedUnit"
 #define D_JSON_SSID "SSId"
 #define D_JSON_STARTDST "StartDST"       // Start Daylight Savings Time
 #define D_JSON_STARTED "Started"
@@ -226,6 +230,7 @@
 #define D_CMND_DRIVER "Driver"
 #define D_CMND_SAVEDATA "SaveData"
 #define D_CMND_SETOPTION "SetOption"
+#define D_CMND_SO "SO"
 #define D_CMND_TEMPERATURE_RESOLUTION "TempRes"
 #define D_CMND_HUMIDITY_RESOLUTION "HumRes"
 #define D_CMND_PRESSURE_RESOLUTION "PressRes"
@@ -287,8 +292,13 @@
 #define D_CMND_LEDSTATE "LedState"
 #define D_CMND_LEDMASK "LedMask"
 #define D_CMND_WIFIPOWER "WifiPower"
+#define D_CMND_SPEEDUNIT "SpeedUnit"
 #define D_CMND_I2CSCAN "I2CScan"
 #define D_CMND_I2CDRIVER "I2CDriver"
+#define D_CMND_DEVGROUP_NAME "DevGroupName"
+#define D_CMND_DEVGROUP_SEND "DevGroupSend"
+#define D_CMND_DEVGROUP_SHARE "DevGroupShare"
+#define D_CMND_DEVGROUPSTATUS "DevGroupStatus"
 #define D_CMND_SERIALSEND "SerialSend"
 #define D_CMND_SERIALDELIMITER "SerialDelimiter"
 #define D_CMND_BAUDRATE "Baudrate"
@@ -299,6 +309,9 @@
   #define D_JSON_FLAG "FLAG"
   #define D_JSON_BASE "BASE"
 #define D_CMND_TEMPOFFSET "TempOffset"
+#define D_CMND_HUMOFFSET "HumOffset"
+#define D_CMND_GLOBAL_TEMP "GlobalTemp"
+#define D_CMND_GLOBAL_HUM "GlobalHum"
 
 // Commands xdrv_01_mqtt.ino
 #define D_CMND_MQTTLOG "MqttLog"
@@ -380,10 +393,12 @@
 #define D_CMND_LED "Led"
 #define D_CMND_LEDTABLE "LedTable"
 #define D_CMND_FADE "Fade"
+#define D_CMND_PALETTE "Palette"
 #define D_CMND_PIXELS "Pixels"
 #define D_CMND_RGBWWTABLE "RGBWWTable"
 #define D_CMND_ROTATION "Rotation"
 #define D_CMND_SCHEME "Scheme"
+#define D_CMND_SEQUENCE_OFFSET "SequenceOffset"
 #define D_CMND_SPEED "Speed"
 #define D_CMND_WAKEUP "Wakeup"
 #define D_CMND_WAKEUPDURATION "WakeUpDuration"
@@ -469,7 +484,6 @@
 
 // Commands xdrv_23_zigbee.ino
 #define D_PRFX_ZB "Zb"
-#define D_PRFX_ZIGBEE "Zigbee"
 #define D_ZIGBEE_NOT_STARTED "Zigbee not started (yet)"
 #define D_CMND_ZIGBEE_PERMITJOIN "PermitJoin"
 #define D_CMND_ZIGBEE_STATUS "Status"
@@ -484,17 +498,38 @@
   #define D_JSON_ZIGBEEZCL_RAW_RECEIVED "ZbZCLRawReceived"
   #define D_JSON_ZIGBEE_DEVICE "Device"
   #define D_JSON_ZIGBEE_NAME "Name"
+  #define D_JSON_ZIGBEE_CONFIRM "ZbConfirm"
 #define D_CMND_ZIGBEE_NAME "Name"
+#define D_CMND_ZIGBEE_MODELID "ModelId"
+  #define D_JSON_ZIGBEE_MODELID "ModelId"
 #define D_CMND_ZIGBEE_PROBE "Probe"
 #define D_CMND_ZIGBEE_FORGET "Forget"
 #define D_CMND_ZIGBEE_SAVE "Save"
   #define D_CMND_ZIGBEE_LINKQUALITY "LinkQuality"
+  #define D_CMND_ZIGBEE_ENDPOINT "Endpoint"
+  #define D_CMND_ZIGBEE_GROUP "Group"
 #define D_CMND_ZIGBEE_READ "Read"
 #define D_CMND_ZIGBEE_SEND "Send"
   #define D_JSON_ZIGBEE_ZCL_SENT "ZbZCLSent"
 #define D_JSON_ZIGBEE_RECEIVED "ZbReceived"
-#define D_JSON_ZIGBEE_RECEIVED_LEGACY "ZigbeeReceived"
 #define D_CMND_ZIGBEE_BIND "Bind"
+  #define D_JSON_ZIGBEE_BIND "ZbBind"
+#define D_CMND_ZIGBEE_UNBIND "Unbind"
+  #define D_JSON_ZIGBEE_UNBIND "ZbUnbind"
+#define D_CMND_ZIGBEE_BIND_STATE "BindState"
+  #define D_JSON_ZIGBEE_BIND_STATE "ZbBindState"
+#define D_CMND_ZIGBEE_PING "Ping"
+  #define D_JSON_ZIGBEE_PING "ZbPing"
+  #define D_JSON_ZIGBEE_IEEE "IEEEAddr"
+#define D_JSON_ZIGBEE_RESPONSE "ZbResponse"
+  #define D_JSON_ZIGBEE_CMD "Command"
+  #define D_JSON_ZIGBEE_STATUS "Status"
+  #define D_JSON_ZIGBEE_STATUS_MSG "StatusMessage"
+#define D_CMND_ZIGBEE_LIGHT "Light"
+  #define D_JSON_ZIGBEE_LIGHT "Light"
+#define D_CMND_ZIGBEE_RESTORE "Restore"
+#define D_CMND_ZIGBEE_CONFIG "Config"
+  #define D_JSON_ZIGBEE_CONFIG "Config"
 
 // Commands xdrv_25_A4988_Stepper.ino
 #define D_CMND_MOTOR "MOTOR"
@@ -509,10 +544,13 @@
 #define D_PRFX_SHUTTER "Shutter"
 #define D_CMND_SHUTTER_OPEN "Open"
 #define D_CMND_SHUTTER_CLOSE "Close"
+#define D_CMND_SHUTTER_TOGGLE "Toggle"
 #define D_CMND_SHUTTER_UP "Up"
 #define D_CMND_SHUTTER_DOWN "Down"
-#define D_CMND_SHUTTER_TOGGLEUP "ToggleUp"
-#define D_CMND_SHUTTER_TOGGLEDOWN "ToggleDown"
+#define D_CMND_SHUTTER_STOPOPEN "StopOpen"
+#define D_CMND_SHUTTER_STOPCLOSE "StopClose"
+#define D_CMND_SHUTTER_STOPTOGGLE "StopToggle"
+#define D_CMND_SHUTTER_STOPPOSITION "StopPosition"
 #define D_CMND_SHUTTER_STOP "Stop"
 #define D_CMND_SHUTTER_POSITION "Position"
 #define D_CMND_SHUTTER_OPENTIME "OpenDuration"
@@ -527,9 +565,19 @@
 #define D_CMND_SHUTTER_BUTTON "Button"
 #define D_CMND_SHUTTER_LOCK "Lock"
 #define D_CMND_SHUTTER_ENABLEENDSTOPTIME "EnableEndStopTime"
+#define D_CMND_SHUTTER_INVERTWEBBUTTONS "InvertWebButtons"
 
 // Commands xdrv_32_hotplug.ino
 #define D_CMND_HOTPLUG "HotPlug"
+
+// Commands xdrv_34_pwm_dimmer.ino
+#ifdef USE_PWM_DIMMER
+#define D_CMND_BRI_PRESET "BriPreset"
+#endif
+
+// Commands xdrv_38_ping.ino
+#define D_CMND_PING "Ping"
+#define D_JSON_PING "Ping"
 
 // Commands xsns_02_analog.ino
 #define D_CMND_ADCPARAM "AdcParam"
@@ -607,7 +655,6 @@ const char S_JSON_DRIVER_INDEX_NVALUE[] PROGMEM =             "{\"" D_CMND_DRIVE
 const char S_JSON_DRIVER_INDEX_SVALUE[] PROGMEM =             "{\"" D_CMND_DRIVER "%d\":\"%s\"}";
 
 const char JSON_SNS_TEMP[] PROGMEM = ",\"%s\":{\"" D_JSON_TEMPERATURE "\":%s}";
-const char JSON_SNS_TEMPHUM[] PROGMEM = ",\"%s\":{\"" D_JSON_TEMPERATURE "\":%s,\"" D_JSON_HUMIDITY "\":%s}";
 
 const char JSON_SNS_ILLUMINANCE[] PROGMEM = ",\"%s\":{\"" D_JSON_ILLUMINANCE "\":%d}";
 const char JSON_SNS_MOISTURE[] PROGMEM = ",\"%s\":{\"" D_JSON_MOISTURE "\":%d}";
@@ -629,11 +676,22 @@ const char S_OFFLINE[] PROGMEM = D_OFFLINE;
 // support.ino
 static const char kMonthNames[] = D_MONTH3LIST;
 
+const char kSpeedUnit[] PROGMEM = "|m/s|km/h|kn|mph|ft/s|yd/s";
+const float kSpeedConversionFactor[] = {1,            // none
+                                        1,            // m/s
+                                        3.6,          // km/h
+                                        1.943844492,  // kn
+                                        2.236936292,  // mph
+                                        3.280839895,  // ft/s
+                                        1.093613298   // yd/s
+                                       };
+
 // xdrv_02_webserver.ino
 #ifdef USE_WEBSERVER
 // {s} = <tr><th>, {m} = </th><td>, {e} = </td></tr>
 const char HTTP_SNS_TEMP[] PROGMEM = "{s}%s " D_TEMPERATURE "{m}%s&deg;%c{e}";
 const char HTTP_SNS_HUM[] PROGMEM = "{s}%s " D_HUMIDITY "{m}%s%%{e}";
+const char HTTP_SNS_DEW[] PROGMEM = "{s}%s " D_DEWPOINT "{m}%s&deg;%c{e}";
 const char HTTP_SNS_PRESSURE[] PROGMEM = "{s}%s " D_PRESSURE "{m}%s %s{e}";
 const char HTTP_SNS_SEAPRESSURE[] PROGMEM = "{s}%s " D_PRESSUREATSEALEVEL "{m}%s %s{e}";
 const char HTTP_SNS_ANALOG[] PROGMEM = "{s}%s " D_ANALOG_INPUT "%d{m}%d{e}";