+++ tasmota/settings.h
@@ -86,7 +86,7 @@ typedef union {                            // Restricted by MISRA-C Rule 18.4 bu
     uint32_t energy_weekend : 1;           // bit 20 (v6.6.0.8)  - CMND_TARIFF
     uint32_t dds2382_model : 1;            // bit 21 (v6.6.0.14) - SetOption71 - Select different Modbus registers for Active Energy (#6531)
     uint32_t hardware_energy_total : 1;    // bit 22 (v6.6.0.15) - SetOption72 - Enable hardware energy total counter as reference (#6561)
-    uint32_t ex_cors_enabled : 1;          // bit 23 (v7.0.0.1)  - SetOption73 - Enable HTTP CORS
+    uint32_t mqtt_buttons : 1;             // bit 23 (v8.2.0.3)  - SetOption73 - Detach buttons from relays and enable MQTT action state for multipress
     uint32_t ds18x20_internal_pullup : 1;  // bit 24 (v7.0.0.1)  - SetOption74 - Enable internal pullup for single DS18x20 sensor
     uint32_t grouptopic_mode : 1;          // bit 25 (v7.0.0.1)  - SetOption75 - GroupTopic replaces %topic% (0) or fixed topic cmnd/grouptopic (1)
     uint32_t bootcount_update : 1;         // bit 26 (v7.0.0.4)  - SetOption76 - Enable incrementing bootcount when deepsleep is enabled
@@ -104,14 +104,14 @@ typedef union {                            // Restricted by MISRA-C Rule 18.4 bu
     uint32_t alexa_ct_range : 1;           // bit 0 (v8.1.0.2)   - SetOption82 - Reduced CT range for Alexa
     uint32_t zigbee_use_names : 1;         // bit 1 (v8.1.0.4)   - SetOption83 - Use FriendlyNames instead of ShortAddresses when possible
     uint32_t awsiot_shadow : 1;            // bit 2 (v8.1.0.5)   - SetOption84 - (AWS IoT) publish MQTT state to a device shadow
-    uint32_t spare03 : 1;
-    uint32_t spare04 : 1;
-    uint32_t spare05 : 1;
-    uint32_t spare06 : 1;
-    uint32_t spare07 : 1;
-    uint32_t spare08 : 1;
-    uint32_t spare09 : 1;
-    uint32_t spare10 : 1;
+    uint32_t device_groups_enabled : 1;    // bit 3 (v8.1.0.9)   - SetOption85 - Enable Device Groups
+    uint32_t led_timeout : 1;              // bit 4 (v8.1.0.9)   - SetOption86 - PWM Dimmer Turn brightness LED's off 5 seconds after last change
+    uint32_t powered_off_led : 1;          // bit 5 (v8.1.0.9)   - SetOption87 - PWM Dimmer Turn red LED on when powered off
+    uint32_t remote_device_mode : 1;       // bit 6 (v8.1.0.9)   - SetOption88 - PWM Dimmer Buttons control remote devices
+    uint32_t zigbee_distinct_topics : 1;   // bit 7 (v8.1.0.10)  - SetOption89 - Distinct MQTT topics per device for Zigbee (#7835)
+    uint32_t only_json_message : 1;        // bit 8 (v8.2.0.3)   - SetOption90 - Disable non-json MQTT response
+    uint32_t fade_at_startup : 1;          // bit 9 (v8.2.0.3)   - SetOption91 - Enable light fading at start/power on
+    uint32_t pwm_ct_mode : 1;              // bit 10 (v8.2.0.4)  - SetOption92 - Set PWM Mode from regular PWM to ColorTemp control (Xiaomi Philips ...)
     uint32_t spare11 : 1;
     uint32_t spare12 : 1;
     uint32_t spare13 : 1;
@@ -132,7 +132,7 @@ typedef union {                            // Restricted by MISRA-C Rule 18.4 bu
     uint32_t spare28 : 1;
     uint32_t spare29 : 1;
     uint32_t spare30 : 1;
-    uint32_t spare31 : 1;
+    uint32_t spare31 : 1;                  // bit 31
   };
 } SysBitfield4;
 
@@ -140,10 +140,8 @@ typedef union {
   uint32_t data;                           // Allow bit manipulation
   struct {
     uint32_t spare00 : 1;
-    uint32_t spare01 : 1;
-    uint32_t spare02 : 1;
-    uint32_t spare03 : 1;
-    uint32_t time_format : 2;              // (v6.6.0.9) - CMND_TIME
+    uint32_t speed_conversion : 3;         // (v8.1.0.10) - Tx2x sensor
+    uint32_t time_format : 2;              // (v6.6.0.9)  - CMND_TIME
     uint32_t calc_resolution : 3;
     uint32_t weight_resolution : 2;
     uint32_t frequency_resolution : 2;
@@ -207,12 +205,35 @@ typedef union {
     uint8_t spare1 : 1;
     uint8_t spare2 : 1;
     uint8_t spare3 : 1;
-    uint8_t spare4 : 1;
-    uint8_t spare5 : 1;
+    uint8_t bh1750_resolution : 2;         // Sensor10 1,2,3
     uint8_t hx711_json_weight_change : 1;  // Sensor34 8,x - Enable JSON message on weight change
     uint8_t mhz19b_abc_disable : 1;        // Disable ABC (Automatic Baseline Correction for MHZ19(B) (0 = Enabled (default), 1 = Disabled with Sensor15 command)
   };
 } SensorCfg1;
+
+typedef union {
+  uint8_t data;
+  struct {
+  uint8_t nf_autotune : 1;            // Autotune the NF Noise Level
+  uint8_t dist_autotune : 1;          // Autotune Disturber on/off
+  uint8_t nf_autotune_both : 1;        // Autotune over both Areas: INDOORS/OUDOORS
+  uint8_t mqtt_only_Light_Event : 1;  // mqtt only if lightning Irq
+  uint8_t spare4 : 1;
+  uint8_t spare5 : 1;
+  uint8_t spare6 : 1;
+  uint8_t spare7 : 1;
+  };
+} As3935IntCfg;
+
+typedef union {
+  uint16_t data;
+  struct {
+  uint16_t nf_autotune_time : 4;            // NF Noise Autotune Time
+  uint16_t dist_autotune_time : 4;          // Disturber Autotune Time
+  uint16_t nf_autotune_min : 4;             // Min Stages
+  uint16_t spare3 : 4;
+  };
+} As3935Param;
 
 typedef struct {
   uint32_t usage1_kWhtotal;
@@ -232,7 +253,7 @@ typedef struct {
 const uint32_t settings_text_size = 699;   // Settings.text_pool[size] = Settings.display_model (2D2) - Settings.text_pool (017)
 const uint8_t MAX_TUYA_FUNCTIONS = 16;
 
-struct SYSCFG {
+struct {
   uint16_t      cfg_holder;                // 000 v6 header
   uint16_t      cfg_size;                  // 002
   unsigned long save_flag;                 // 004
@@ -243,7 +264,7 @@ struct SYSCFG {
   int16_t       save_data;                 // 014
   int8_t        timezone;                  // 016
 
-  // Start of char array storing all parameter strings
+  // Start of char array storing all parameter strings ********
 
   char          text_pool[101];            // 017 - was ota_url[101] - size is settings_text_size
 
@@ -279,7 +300,7 @@ struct SYSCFG {
   char          ex_button_topic[33];       // 290
   char          ex_mqtt_grptopic[33];      // 2B1
 
-  // End of single char array of 698 chars max
+  // End of single char array of 698 chars max ****************
 
   uint8_t       display_model;             // 2D2
   uint8_t       display_mode;              // 2D3
@@ -337,8 +358,17 @@ struct SYSCFG {
   SysBitfield3  flag3;                     // 3A0
   uint8_t       switchmode[MAX_SWITCHES];  // 3A4  (6.0.0b - moved from 0x4CA)
 
+#ifdef ESP8266
   char          ex_friendlyname[4][33];    // 3AC
   char          ex_switch_topic[33];       // 430
+#else  // ESP32
+  myio          my_gp;                     // 3AC - 2 x 40 bytes (ESP32)
+  mytmplt       user_template;             // 3FC - 2 x 37 bytes (ESP32)
+
+  uint8_t       free_esp32_446[10];        // 446
+
+  uint8_t       esp32_webcam_resolution;   // 450
+#endif  // ESP8266 - ESP32
 
   char          serial_delimiter;          // 451
   uint8_t       seriallog_level;           // 452
@@ -348,7 +378,13 @@ struct SYSCFG {
   uint8_t       module;                    // 474
   uint8_t       ws_color[4][3];            // 475
   uint8_t       ws_width[3];               // 481
-  myio          my_gp;                     // 484
+
+#ifdef ESP8266
+  myio          my_gp;                     // 484 - 17 bytes (ESP8266)
+#else  // ESP32
+  uint8_t       free_esp32_484[17];        // 484
+#endif  // ESP8266 - ESP32
+
   uint8_t       my_adc0;                   // 495
   uint16_t      light_pixels;              // 496
   uint8_t       light_color[5];            // 498
@@ -399,7 +435,15 @@ struct SYSCFG {
   uint16_t      mcp230xx_int_timer;        // 718
   uint8_t       rgbwwTable[5];             // 71A
   uint8_t       user_template_base;        // 71F
-  mytmplt       user_template;             // 720  29 bytes
+
+  char          user_template_name[15];    // 720  15 bytes - Backward compatibility since v8.2.0.3
+
+#ifdef ESP8266
+  mytmplt       user_template;             // 72F  14 bytes (ESP8266)
+#else  // ESP32
+  uint8_t       free_esp32_72f[14];        // 72F
+#endif  // ESP8266 - ESP32
+
   uint8_t       novasds_startingoffset;    // 73D
   uint8_t       web_color[18][3];          // 73E
   uint16_t      display_width;             // 774
@@ -465,10 +509,36 @@ struct SYSCFG {
   uint8_t       mqttlog_level;             // F01
   uint8_t       sps30_inuse_hours;         // F02
   uint8_t       hotplug_scan;              // F03
-  uint8_t       reserved1;                 // F04 - reserved for s-hadinger
+  uint8_t       bri_power_on;              // F04
+  uint8_t       bri_min;                   // F05
+  uint8_t       bri_preset_low;            // F06
+  uint8_t       bri_preset_high;           // F07
+  int8_t        hum_comp;                  // F08
+  uint8_t       wifi_channel;              // F09
+  uint8_t       wifi_bssid[6];             // F0A
+  uint8_t       as3935_sensor_cfg[5];      // F10
+  As3935IntCfg  as3935_functions;          // F15
+  As3935Param   as3935_parameter;          // F16
+  uint64_t      zb_ext_panid;              // F18
+  uint64_t      zb_precfgkey_l;            // F20
+  uint64_t      zb_precfgkey_h;            // F28
+  uint16_t      zb_pan_id;                 // F30
+  uint8_t       zb_channel;                // F32
+  uint8_t       zb_free_byte;              // F33
+  uint16_t      pms_wake_interval;         // F34
+  uint8_t       config_version;            // F36
 
-  uint8_t       free_f05[207];             // F05
+  uint8_t       free_f37[129];             // F37 - Decrement if adding new Setting variables just above and below
 
+  // Only 32 bit boundary variables below
+  uint16_t      pulse_counter_debounce_low;  // FB8
+  uint16_t      pulse_counter_debounce_high; // FBA
+  uint32_t      keeloq_master_msb;         // FBC
+  uint32_t      keeloq_master_lsb;         // FC0
+  uint32_t      keeloq_serial;             // FC4
+  uint32_t      keeloq_count;              // FC8
+  uint32_t      device_group_share_in;     // FCC - Bitmask of device group items imported
+  uint32_t      device_group_share_out;    // FD0 - Bitmask of device group items exported
   uint32_t      bootcount_reset_time;      // FD4
   int           adc_param4;                // FD8
   uint32_t      shutter_button[MAX_KEYS];  // FDC
@@ -477,13 +547,17 @@ struct SYSCFG {
   uint32_t      cfg_crc32;                 // FFC
 } Settings;
 
-struct RTCRBT {
+typedef struct {
   uint16_t      valid;                     // 280 (RTC memory offset 100 - sizeof(RTCRBT))
   uint8_t       fast_reboot_count;         // 282
   uint8_t       free_003[1];               // 283
-} RtcReboot;
+} TRtcReboot;
+TRtcReboot RtcReboot;
+#ifdef ESP32
+RTC_NOINIT_ATTR TRtcReboot RtcDataReboot;
+#endif
 
-struct RTCMEM {
+typedef struct {
   uint16_t      valid;                     // 290 (RTC memory offset 100)
   uint8_t       oswatch_blocked_loop;      // 292
   uint8_t       ota_loader;                // 293
@@ -499,7 +573,11 @@ struct RTCMEM {
 
   uint8_t       free_022[22];              // 2D6
                                            // 2EC - 2FF free locations
-} RtcSettings;
+} TRtcSettings;
+TRtcSettings RtcSettings;
+#ifdef ESP32
+RTC_NOINIT_ATTR TRtcSettings RtcDataSettings;
+#endif
 
 struct TIME_T {
   uint8_t       second;